requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
dj-database-url==0.5.0
Django==3.2.10
django-cors-headers==3.10.1
django-heroku==0.3.1
django-ranged-response==0.2.0
django-recaptcha==2.0.6
django-simple-captcha==0.5.14
django-utils-six==2.0
docopt==0.6.2
gunicorn==20.1.0
idna==3.3
Pillow==8.4.0
pipreqs==0.4.11
psycopg2==2.9.2
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
yarg==0.1.9
